add scalar transforms
- scale transform
- rename outputs from "primary" to "output"
- change data type from double to float

+++ engine/src/components/hsv_component.h
@@ -10,8 +10,8 @@ public:
 
   Color calculate(const FrameContext& frame) const override;
   
-  SignalFunction<double> hue;
-  SignalFunction<double> saturation;
-  SignalFunction<double> value;
+  SignalFunction<float> hue;
+  SignalFunction<float> saturation;
+  SignalFunction<float> value;
 
 };
