+++ engine/src/components/oscillators/saw_wave.cpp
@@ -1,7 +1,7 @@
 #include "saw_wave.h"
 
-double SawWave::calculate(const FrameContext& frame) const
+float SawWave::calculate(const FrameContext& frame) const
 {
-  double unused;
+  float unused;
   return amplitude(frame) * modf(frame.timeSeconds() * frequency(frame) + phase(frame), &unused);
 }
