+++ engine/src/components/oscillators/saw_wave.h
@@ -6,6 +6,6 @@
 class SawWave : public Waveform {
 
 public:
-  double calculate(const FrameContext& frame) const override;
+  float calculate(const FrameContext& frame) const override;
 
 };
