+++ engine/src/components/oscillators/square_wave.h
@@ -6,6 +6,6 @@
 class SquareWave : public Waveform {
 
 public:
-  double calculate(const FrameContext& frame) const override;
+  float calculate(const FrameContext& frame) const override;
 
 };
