+++ engine/src/components/pixel_x_component.h
@@ -3,10 +3,10 @@
 
 #include "basic_component.h"
 
-class PixelXComponent : public BasicComponent<double> {
+class PixelXComponent : public BasicComponent<float> {
 
 public:
 
-  double calculate(const FrameContext& frame) const override;
+  float calculate(const FrameContext& frame) const override;
 
 };
